Quiz Wiz Pro - v1.3
Added Start Menu, there the program starts and gives the user the option to use as a Admin/Teacher or Student.  Added Login Menu For Student and Admins. The username and password for the Student account is "student" and password. The user name and password for the admin is
"admin" and "password.