Uploading the GUI for the menu
Added the GUIs and everything should be fine! I've ran it several times on my end. I just don't have a local instance of the database to test DB Connection and logins and creating users. However, I used the same methods you all used and it should work the same :)